+++ packages/engine/src/experiment/error.rs
@@ -12,6 +12,9 @@ pub enum Error {
     #[error("{0}")]
     Unique(String),
 
+    #[error("Execution error: {0}")]
+    Execution(#[from] execution::Error),
+
     #[error("Controller error: {0}")]
     Controller(#[from] ControllerError),
