+++ packages/engine/src/experiment/package/simple.rs
@@ -1,14 +1,13 @@
 use std::{collections::HashMap, sync::Arc};
 
-use execution::package::experiment::comms::{update::ExpPkgUpdateRecv, ExperimentControl};
+use execution::package::experiment::comms::{
+    control::ExpPkgCtlSend, update::ExpPkgUpdateRecv, ExperimentControl,
+};
 use simulation_structure::SimulationShortId;
 
 use crate::{
     config::ExperimentConfig,
-    experiment::{
-        controller::comms::exp_pkg_ctl::ExpPkgCtlSend,
-        error::{Error, Result},
-    },
+    experiment::error::{Error, Result},
     proto::SimpleExperimentConfig,
 };
